+++ cli/command/trust/key_generate_test.go
@@ -130,7 +130,7 @@ func TestValidateKeyArgs(t *testing.T) {
 	assert.NoError(t, ioutil.WriteFile(filepath.Join(pubKeyCWD, "a.pub"), []byte("abc"), notary.PrivExecPerms))
 	err = validateKeyArgs("a", pubKeyCWD)
 	assert.Error(t, err)
-	assert.Equal(t, err.Error(), fmt.Sprintf("public key file already exists: \"%s/a.pub\"", pubKeyCWD))
+	assert.Equal(t, err.Error(), fmt.Sprintf("public key file already exists: \"%s\"", filepath.Join(pubKeyCWD, "a.pub")))
 
 	err = validateKeyArgs("a", "/random/dir/")
 	assert.Error(t, err)
